KAFKA-18199: Fix size calculation for nullable tagged structs
When a struct field is tagged and nullable, it is serialized as
{ varint tag; varint dataLength; nullable<struct> data }, where
nullable<struct> is serialized as
{ varint isNotNull; if (isNotNull) struct s; }. The length field
includes the is-not-null varint.

This patch fixes a bug in serialization where the written value of
the length field and the value used to compute the size of the length
field differed by 1. In practice this had no impact unless the
serialized length of the struct is 127 bytes, since the varint encodings
of 127 and 128 have different lengths (7f vs 80 01).