KAFKA-18284: Add group coordinator records for Streams rebalance protocol
KIP-1071 specifies records that the group coordinator needs to store
into the consumer offset topic to persist the state of a Streams
group. This records are specified in json files from which the actual
classes for the records are generated.

This commit adds the records needed by the group corrdinator to store
the state of a Streams group.