Address some comments
Turn timeSinceLastPollMs into a local

Add some comments to recordPollStart

Make recordPollStart handle polling intervals of 0 ms by ignoring them,
rather than letting the metric value become NaN

Try to improve the explanation in KafkaConsumerTest of how this metric is
intended to work, and why it excludes the last poll.

Move time increments in the test around a bit. A polling interval
runs from poll start to poll start, which means that the time
in an interval we consider to be "outside of poll" is the time after
the poll completes, until the next poll starts. This means the test
is easier to read if we increment the time after each poll,
rather than before each poll.

Add bit to the test that verifies the 0ms interval handling correctly ignores
such intervals.